backend/requirements.txt
# ISPL Insurance Policy AI Backend Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# LangChain & LangGraph
langchain==0.1.0
langgraph==0.0.66
langchain-openai==0.0.5

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
tabula-py==2.9.0

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# NLP & Embeddings
openai==1.3.7
tiktoken==0.5.2

# Additional dependencies for multi-agent system
numpy==1.24.3
pandas==2.1.3

# Vector Database
# pgvector==0.2.4

# Caching
# redis==5.0.1
